recommendation-service/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.8.1
scikit-surprise==1.1.1
kafka-python==2.0.2

#Testing
pandas==1.3.5
tqdm==4.63.0
pytest==7.1.3
pytest-html==3.1.1
coverage==6.3.2
numpy==1.22.3
requests==2.28.1

#Deployment
waitress==2.1.2

#Provenance
boto3==1.16.0
influxdb==5.3.1
neptune-client==0.16.13
